swap intro with get-started (#3406)
* swap intro with get-started

* fix routing
