+++ services/ui-src/src/forms/mcpar/index.ts
@@ -76,11 +76,7 @@ export const mcparStructureWithData = addDataToReportStructure(
 
 export const mcparRoutes = makeRouteArray(mcparStructureWithData);
 
-export const nonReportMcparRoutes = [
-  "/mcpar/intro",
-  "/mcpar/dashboard",
-  "/mcpar/getting-started",
-];
+export const nonReportMcparRoutes = ["/mcpar/dashboard", "/mcpar/get-started"];
 
 export const isMcparReportPage = (pathname: string) => {
   return (
